replication-canary consumes mysql port from the mysql bosh link
Previously replication canary required that the mysql port be specified
or it would default to port 3306.  Now it automatically reads the mysql
port from the 'mysql' bosh link when one exists.  If the mysql link does
not exist or is explicitly disabled in a manifest, replication-canary
will fallback to using the release properties.

[#164440355]

Signed-off-by: Andrew Garner <agarner@pivotal.io>