Sum wsrep_cluster_status and include only on HA
For 3-node galera clusters, we want to warn if fewer than three nodes
report good status and critical alert if fewer than two nodes report good
status.

This part of the yaml is omitted entirely for smaller deployments, since
the metric "wsrep_cluster_status" is not applicable unless there is a
cluster.

[#167328597]

Signed-off-by: Alvaro Perez-Shirley <aperezshirley@pivotal.io>