replication-canary: use bpm for replication-canary job
Previously replication-canary was started via a shell script that ran
the replication-canary binary as the OS "root" user.

Now bpm is used to start replication-canary so that it runs under the
standard BOSH "vcap" user.

By default, bpm processes will not have access to /var/vcap/sys/run/ so
the pidfile management previously done by replication-canary has been
removed.  bpm will manage pidfiles directly so this does not need to be
implemented in the utility.

[#182727732](https://www.pivotaltracker.com/story/show/182727732)