operations: ensure loggregator-agent source release is uploaded
Update the pxc-add-metrics.yml ops-file to fetch the latest
loggregator-agent source release from bosh io.

This ensures that we can always deploy against the desired stemcell even
if the bosh environment only has a compiled loggregator-agent-release.
This might happen if a source release for loggregator-agent-release has
never been uploaded, but cf was deployed with the compiled-releases
ops-file.

[#183002706](https://www.pivotaltracker.com/story/show/183002706)

Co-authored-by: Andrew Garner <garnera@vmware.com>
Co-authored-by: Kim Bassett <kbassett@vmware.com>