mysql-metrics: add mysql-metrics.port property for configuring the mysql port
In some environments we would like to configure the MySQL port when it
is set to a non-standard value.

Rewrote the spec / templating test script to use the bosh template gem
rather than the bosh-template cli.  The former handles reading default
values from the job spec whereas "bosh-template" is glorified erb
renderer and ignores defaults in the job spec.

[#183015776](https://www.pivotaltracker.com/story/show/183015776)

Co-authored-by: Andrew Garner <garnera@vmware.com>
Co-authored-by: Kim Bassett <kbassett@vmware.com>