mysql-metrics: integration-test uses docker for mysql testing
Previously the integration test required an external MySQL to be spun up
in order for tests to pass and the user must specify configuration to
connect to the MySQL.

Now the test uses docker to create a percona-server instance that's used
for integration testing.  This is an integration test to demonstrate we
can connect to an arbitrary MySQL port rather than the previous behavior
where mysql-metrics hardcoded port 3306.

[#183015776](https://www.pivotaltracker.com/story/show/183015776)

Co-authored-by: Andrew Garner <garnera@vmware.com>
Co-authored-by: Kim Bassett <kbassett@vmware.com>