mysql-diag: decouple replication-canary from mysql-diag
With this change, replication canary is an _optional_ component of
mysql-diag.  When present mysql-diag will work precisely as it does
today.

Without replication-canary, mysql-diag MUST be configured with explicit
database credentials.

When mysql-diag is configured with explicit database credentials AND
alongside replication-canary, the explicit credentials are preferred
and the replication-canary credentials ignored.

Notable changes:

- add bosh template test for mysql-diag/config/mysql-diag-config.yml.erb
- conditionally consume the replication-canary link for backwards
  compatibility.
- mysql-diag now supports two new spec properties: db_{username,password}

[#186704976](https://www.pivotaltracker.com/story/show/186704976)

Co-authored-by: Kim Bassett <kbassett@vmware.com>
Co-authored-by: Andrew Garner <garnera@vmware.com>