WIP - log new disk I/O metric
Brings in prometheus procfs library to try and replicate what the node
exporter does in its diskstats collector

This results in a new log being emitted when running:
`cf tail -f p-mysql`

But still need to figure out exactly which disk(s) to monitor for the
metric. A comment is added to show the output of `df -h` and another
comment shows all the different Diskstats that are available

As a design observation - the current pattern in the processor of
needing to have both a "gatherer" and a "metricsComputer" feels poor. In
general, we don't want to have one dependency to get data and another to
act on it - we just want one dependency that does the thing. Not sure if
this story will be the best place to review this, but it feels like
something new should be added that has both responsibilities. I could
imagine the processor ends up turning into something like a pipeline,
with each component doing some work in the larger assembly line.

[TNZ-22465](https://jira.eng.vmware.com/browse/TNZ-22465)

Authored-by: Ryan Wittrup <ryan.wittrup@broadcom.com>