Refactor: trying out pipeline design
First in a phase of refactorings to better isolate the metrics by type.
Each calculator should have full control over how it gets the data it
needs to calculate the metrics it owns.

The result is a pipeline, where adding a new metrics means defining a
new Calculator and adding it to the pipeline.

Authored-by: Ryan Wittrup <ryan.wittrup@broadcom.com>