Refactors acceptance tests to avoid flakes
While testing in the new shepherd v1 environment, the ephemeral disk
test began to fail - not sure how it worked before, but it was using `cf
tail` to get the metric value. But since it was running an open `cf
tail` without any filtering, and it always took the first value it
found, there was no guarantee that the first metric it found was emitted
from mysql/0

Updates the test to find the VM's ID via bosh, and then uses that ID to
run `cf query` instead - this allows us to select for the correct ID to
make sure we are always getting the metrics from mysql/0

This change also consolidates what was two different "By" blocks.
Previously we would get the disk usage after allocating data in one
block, and then assert on the delta in another block. That felt overly
complex, and we can instead combine this via the eventually block to
directly assert the value is within our expected range

Minor refactoring to use a more standard pattern for running bosh
commands

[TNZ-37039](https://vmw-jira.broadcom.net/browse/TNZ-37039)

Authored-by: Ryan Wittrup <ryan.wittrup@broadcom.com>