Replace lager usage with log/slog
slog is faster and simpler and reduces the amount of data logged while
improving the clarity of the logs.

This removes an external third party dependency chain and simply uses
the Go stdlib log/slog for logging.

The old log format was maintained as much as possible by using
slog.ReplaceAttr and maintaing the legacy nested "data.metric" structure
of the existing package. However the "source" key has been removed -
this always had a static "MetricsEmitter" value that seemed low value
and whose removal would not be missed.