Fix mysql-diag test
Last test was an incomplete update to assert mysql-diag behavior.

Now mysql-diag is run in a loop while running a pxc-release bootstrap
operation to assert that - with changes in the galera-agent component in
pxc-release - mysql-diag does not interfer with a concurrent bootstrap
operation.

Cleaned up and removed the legacy testhelpers package leveraging more
modern test helpers in the utilities package.